Move `BulkIndexFailure` into namespace
This exception is better at home in the `Indexer` namespace.
