+++ lib/indexer.rb
@@ -1,6 +1,7 @@
 require 'indexer/popularity_lookup'
 require 'indexer/bulk_payload_generator'
 require 'indexer/document_preparer'
+require 'indexer/exceptions'
 require 'indexer/tag_lookup'
 
 module Indexer
