+++ test/unit/elasticsearch_index_test.rb
@@ -108,7 +108,7 @@ class ElasticsearchIndexTest < MiniTest::Unit::TestCase
     begin
       @index.add(documents)
       flunk("No exception raised")
-    rescue SearchIndices::BulkIndexFailure => e
+    rescue Indexer::BulkIndexFailure => e
       assert_equal "Failed inserts: /foo/baz (stuff)", e.message
       assert_equal ["/foo/baz"], e.failed_keys
     end
